need to fix lobby.
need to fix leaving lobby, players object destroyed and username goes to placeholder with all data removed. when inviting a player the invited players name is a bunch of random characters instead of their username. when joining the host player from the invite the player apears on the host side but the invited player still only sees their own personal lobby and is alone.