fix cosmetics and player linking
Player Cosmetic script isnt linked to cosmetics. when invited or host player click leave only the other sees the person leave, while the one that clicked leave still sees more then 1 player character in the lobby. and when host kicks an player , the hosts lobby shows that person being removed but the kicked person still sees the host / others in the lobby. but the amount of people shown in the lobby does update showing there is 1 player even tho there is more than 1 player model.